+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 - Fixed CommandLine Parser not being disposed after argument parsing
 ### Changed
 - SDK - Updated DotNet SDK to 10.0.300
+- Updated FunFair.Test.Common to 6.2.25.2243 in all test projects
 ### Deprecated
 ### Removed
 ### Deployment Changes